requirements.txt
railtracks>=1.3.1
anthropic>=0.20.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0
